test: move mstsgu HTTP auth tests off the lib crate
xtask test-settings requires [lib] test = false on every crate Cargo.toml. Lib-crate unit tests therefore never run.

Move the http_auth cases to tests/http_auth.rs, expose the helpers through a doc-hidden public module, and run that target from xtask with native-tls so workspace tests still compile without a TLS backend.
